Movie on mobile prototype made using proto.io and Invision
General Objectives:

Work with users determine requirements for your system
Design the interface
Prototype the interface
Evaluate the Interface
Problem: Media and Stress

Too often, the news, social media and other outlets filled with
negative reports that affect not only those directly involved, but also
those who read about them[1]. There are not a lot of mechanisms that
support the reporting of good events, random acts of kindness, or
courageous behaviors. For this project, you must design an app that
supports the reporting and acknowledgement of the positive side of
human behavior OR an app that presents balanced types of news on UF’s
campus involving faculty, staff, and/or students.

Notes:

This must be actively engaging, a blog or other simple news reporting
system is not sufficient.
You might consider moving beyond just reporting, to creating a
mechanism that actually supports the carrying out of random acts of
kindness or other prosocial behaviors.
The application does not have to be restricted to the phone/tablet.
Requirements

Create a question guide to support your interviews.
Interview two (Graduate Students: four) potential users of this app.
Take notes or you might want to record the conversation (Be sure to ask
permission first).
Based on your interviews, sketch two potential designs for you app
using inVision. (http://www.invisionapp.com/)
Once you have completed the sketches, share your sketches with two
(Graduate students: four) potential users and have them comment on your
designs.
Choose ONE of your sketches as the final prototype to use for our in
class evaluation.
On January 27, in class have one sketch selected to share. During class
you will have the opportunity to share your app with classmates and
receive feedback.
(1-10, 10 being the highest) Would you recommend our app to a friend?
What did you like least about the app?
How easy is the app to use
How easy is the app to use?
What features are of the app are the most important to you?
What features of our app are the least important to you?
Create a video report of your process. This video should adhere to the
following guidelines:
Less that 5 minutes runtime [10 points]
Provide an example of how your app would function. [20 points]
Demonstrate two tasks that can be accomplished within the app (i.e.,
show how a user could navigate through the app to accomplish a goal).
[40 points]
Include a summary of feedback from the in class exercise. [30 points]